+++ lib/flutter_context_menu.dart
@@ -1,9 +1,13 @@
+export 'src/components/checkable_menu_item.dart';
 export 'src/components/menu_divider.dart';
+export 'src/core/models/checkable_controller.dart';
 export 'src/components/menu_header.dart';
 export 'src/components/menu_item.dart';
 export 'src/core/enums/spawn_direction.dart';
 export 'src/core/models/context_menu.dart';
+export 'src/core/models/context_menu_checkable_item.dart';
 export 'src/core/models/context_menu_entry.dart';
+export 'src/core/models/context_menu_interactive_entry.dart';
 export 'src/core/models/context_menu_item.dart';
 export 'src/core/utils/extensions/single_activator_ext.dart';
 export 'src/core/utils/helpers.dart';